+++ packages/gamut/src/index.tsx
@@ -1,3 +1,4 @@
+export * from './Alert';
 export * from './AccordionButton';
 export * from './AppBar';
 export * from './AppBar/AppBarSection';
@@ -23,7 +24,6 @@ export * from './Icon';
 export * from './IkonaIcon';
 export * from './Interstitial';
 export * from './Layout';
-export * from './Logo';
 export * from './Markdown';
 export * from './Menus/MenuItem';
 export * from './Menus/SideMenu';