+++ packages/styleguide/stories/Atoms/ProgressBar/ProgressBar.stories.tsx
@@ -7,22 +7,23 @@ import styles from './styles.module.scss';
 import {
   StoryTemplate,
   StoryStatus,
-  decoratedStory,
+  decoratedStories,
   StoryDescription,
+  decoratedStory,
 } from '../../Templating';
 
-export default decoratedStory('Atoms', ProgressBar);
+export default decoratedStories('Atoms', ProgressBar);
 
 const bars = [
   {
-    displayLabel: false,
+    large: false,
     style: {
       backgroundColor: colors.blue[100],
       barColor: colors.blue[700],
     },
   },
   {
-    displayLabel: true,
+    large: true,
     style: {
       backgroundColor: colors.gray[100],
       barColor: colors.yellow[500],
@@ -31,7 +32,7 @@ const bars = [
   },
 ];
 
-export const progressBar = () => (
+export const progressBar = decoratedStory(() => (
   <StoryTemplate status={StoryStatus.InProgress}>
     <StoryDescription>
       ProgressBars are to be used when you would like to indicate a known or
@@ -39,17 +40,17 @@ export const progressBar = () => (
       might show one on a quiz page indicating how many questions have been
       completed.
       <br />
-      Bars that <code>displayLabel</code> are thicker, and will display a
+      Bars with <code>large</code> specified are thicker, and will display a
       percentage label if a font color is specified.
     </StoryDescription>
     <LayoutGrid className={styles.progressBarGrid} columnGap="sm" rowGap="sm">
       {[0, 25, 50, 75, 100].map(percent =>
-        bars.map(({ displayLabel, style }) => (
-          <Column key={[percent, displayLabel, style].join()} size={6}>
-            <ProgressBar percent={percent} style={style} />
+        bars.map(({ large, style }) => (
+          <Column key={[percent, large, style].join()} size={6}>
+            <ProgressBar large={large} percent={percent} style={style} />
           </Column>
         ))
       )}
     </LayoutGrid>
   </StoryTemplate>
-);
+));