+++ packages/styleguide/stories/Atoms/Spinner.stories.tsx
@@ -2,15 +2,16 @@ import React from 'react';
 import { Spinner } from '@codecademy/gamut/src';
 import { text } from '@storybook/addon-knobs';
 import {
+  decoratedStories,
   decoratedStory,
   StoryStatus,
   StoryTemplate,
   StoryDescription,
 } from '../Templating';
 
-export default decoratedStory('Atoms', Spinner);
+export default decoratedStories('Atoms', Spinner);
 
-export const spinner = () => (
+export const spinner = decoratedStory(() => (
   <StoryTemplate status={StoryStatus.Ready}>
     <StoryDescription>
       Perpetually spinning circle, such as for a loading indicator. Similar to a{' '}
@@ -18,4 +19,4 @@ export const spinner = () => (
     </StoryDescription>
     <Spinner size={text('size', '5em')} />
   </StoryTemplate>
-);
+));